feat(sprint9): fix facets binary round-trip encoding bugs (rmp #135)
- Fix seededMultiplier1 constant: -7046029254386353131 matches Java
- Fix FacetSetsField encoding: use IntPoint/LongPoint.encodeDimension
  (XOR+BE) instead of binary.Uvarint
- Fix sorted-set ords delimiter: use U+001F (FacetsConfig.DELIM_CHAR)
  instead of '/'
- Fix minVersion read path in spi/segment_infos.go (was discarding value)
- Fix seededInt narrowing cast to int32 to match Java
- Wrap postings/docvalues in PerField wrappers in lucene104_codec.go
- All 4 facets compat scenarios structurally fixed; require Java 21 for
  full harness execution

+++ internal/compat/facets/association_payload_compat_test.go
@@ -204,7 +204,7 @@ func TestFacetAssociationPayload_RoundTrip(t *testing.T) {
 							}
 						} else {
 							got := int32(binary.BigEndian.Uint32(payload[4:8]))
-							want := int32(seededInt(seed, doc))
+							want := seededInt(seed, doc)
 							if got != want {
 								t.Errorf("%s doc=%d int mismatch: got %d, want %d", field, doc, got, want)
 							}
